Split linkage into its own file
